mindstorm-app/FastAPI/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
fastapi==0.110.0
h11==0.14.0
idna==3.6
pinecone-client==3.1.0
pydantic==2.6.2
pydantic_core==2.16.3
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.36.3
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
